+++ src/main/java/domain/card/trump/Rank.java
@@ -1,4 +1,4 @@
-package domain.card;
+package domain.card.trump;
 
 public enum Rank {
