+++ src/main/java/domain/card/trump/Suit.java
@@ -1,4 +1,4 @@
-package domain.card;
+package domain.card.trump;
 
 public enum Suit {
