+++ src/test/java/domain/card/CardTest.java
@@ -1,5 +1,7 @@
 package domain.card;
 
+import domain.card.trump.Rank;
+import domain.card.trump.Suit;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
